+++ src/RootBrowse/test/__init__.py
@@ -1,1 +1,0 @@
-"""RootBrowse 测试包"""